AI_hackathon/requirements.txt
Markdown==3.5.2
markdown_pdf==1.0
mediapipe==0.9.1.0
numpy==1.25.0
openai==1.9.0
opencv-python-headless<4.3
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.2.0
pydub==0.25.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.30.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.12
streamlit_webrtc==0.47.1
Werkzeug==3.0.1
